first test ok with dependencies (remote & downloaded)

+++ src/main/java/net_alchim31_runner/CompilerService.java
@@ -3,17 +3,14 @@ package net_alchim31_runner;
 import java.io.File;
 import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.Collections;
 import java.util.List;
 import java.util.jar.Manifest;
 
-import javax.tools.DiagnosticCollector;
-import javax.tools.JavaCompiler;
-import javax.tools.JavaFileObject;
-import javax.tools.StandardJavaFileManager;
-import javax.tools.ToolProvider;
+import javax.annotation.processing.Processor;
+import javax.tools.StandardLocation;
 
-import org.codehaus.plexus.util.StringUtils;
-
+//TODO may be use plexus-compiler-javac / plexus-compiler-manager / plexus-compiler-api
 public class CompilerService {
   public CompilerService() {
   }
@@ -22,37 +19,69 @@ public class CompilerService {
     return f.getName().endsWith(".java");
   }
   
-  void compileToJar(File dest, File src, List<File> classpath, List<String> options) throws Exception {
+  boolean compileToJar(File dest, File src, List<File> classpath, List<String> options) throws Exception {
     File dir = new File(dest.getAbsolutePath() + ".d");
     if (dir.isDirectory()) {
       FileUtils.cleanDirectory(dir);
     } else {
       dir.mkdirs();
     }
-    compileToDir(dir, src, classpath, options);
-    FileUtils.jar(dest, dir, new Manifest());
-    FileUtils.deleteDirectory(dir);
+    boolean b = compileToDir(dir, src, classpath, options);
+    if (b) {
+      FileUtils.jar(dest, dir, new Manifest());
+      FileUtils.deleteDirectory(dir);
+    }
+    return b;
   }
   
   //TODO add reporter of error, ...
-  void compileToDir(File dest, File src, List<File> classpath, List<String> options) throws Exception {
-    JavaCompiler compiler = ToolProvider.getSystemJavaCompiler();
-    DiagnosticCollector<JavaFileObject> diagnostics = new DiagnosticCollector<JavaFileObject>();
-    StandardJavaFileManager fileManager = compiler.getStandardFileManager(diagnostics, null, null);
+  //TODO provide src as String because it's already read and in memory (to extract classpath,...)
+  //  see http://www.accordess.com/wpblog/an-overview-of-java-compilation-api-jsr-199/
+  //TODO find how to run in isolated classloader
+  boolean compileToDir(File dest, File src, List<File> classpath, List<String> options) throws Exception {
+//    ClassWorld world = new ClassWorld();
+//    ClassRealm javacRealm = world.newRealm("javac");
+//    //javacRealm.addConstituent( containerJarUrl );
+//    System.err.println("javac0 : " + Arrays.toString(javacRealm.getURLs()));
+    final ClassLoader cl0 = Thread.currentThread().getContextClassLoader();
     try {
-      Iterable<? extends JavaFileObject> compilationUnits = fileManager.getJavaFileObjectsFromFiles(Arrays.asList(src));
-      List<String> optionsL = new ArrayList<>(options.size() + 4);
-      optionsL.addAll(options);
-      optionsL.add("-d");
-      optionsL.add(dest.getAbsolutePath());
-      if (classpath.size() > 0) {
-        optionsL.add("-classpath");
-        optionsL.add(StringUtils.join(classpath.iterator(), File.pathSeparator));
+//      Class<javax.tools.ToolProvider> containerClass = javacRealm.loadClass( "javax.tools.ToolProvider" );
+//      Thread.currentThread().setContextClassLoader( javacRealm.getSystemClassLoader() );
+//    javax.tools.JavaCompiler compiler = (javax.tools.JavaCompiler) containerClass.getMethod("getSystemJavaCompiler").invoke(null);
+      Thread.currentThread().setContextClassLoader(javax.tools.ToolProvider.getSystemToolClassLoader());
+      javax.tools.JavaCompiler compiler = javax.tools.ToolProvider.getSystemJavaCompiler();
+      javax.tools.DiagnosticCollector<javax.tools.JavaFileObject> diagnosticsCollector = new javax.tools.DiagnosticCollector<javax.tools.JavaFileObject>();
+      javax.tools.StandardJavaFileManager fileManager = compiler.getStandardFileManager(diagnosticsCollector, null, null);
+      fileManager.setLocation(StandardLocation.CLASS_PATH, classpath);
+      fileManager.setLocation(StandardLocation.CLASS_OUTPUT, Collections.singleton(dest));
+      fileManager.setLocation(StandardLocation.ANNOTATION_PROCESSOR_PATH, Collections.<File>emptyList());
+      fileManager.setLocation(StandardLocation.SOURCE_PATH, Collections.<File>emptyList());
+//      fileManager.setLocation(StandardLocation.SOURCE_OUTPUT, Collections.EMPTY_LIST);
+      try {
+        Iterable<? extends javax.tools.JavaFileObject> compilationUnits = fileManager.getJavaFileObjectsFromFiles(Arrays.asList(src));
+        List<String> optionsL = new ArrayList<>(options.size() + 4);
+        optionsL.addAll(options);
+//        optionsL.add("-d");
+//        optionsL.add(dest.getAbsolutePath());
+//        if (classpath.size() > 0) {
+//          optionsL.add("-classpath");
+//          optionsL.add(StringUtils.join(classpath.iterator(), File.pathSeparator));
+//        }
+        javax.tools.JavaCompiler.CompilationTask task = compiler.getTask(null, fileManager, diagnosticsCollector, optionsL, null, compilationUnits);
+        task.setProcessors(Collections.<Processor>emptyList());
+        boolean b = task.call();
+        
+        for(javax.tools.Diagnostic<? extends javax.tools.JavaFileObject> d : diagnosticsCollector.getDiagnostics()){
+            // Print all the information here.
+            System.err.format("%s:%s:%d:%d:%s:%s\n", d.getCode(), d.getKind(), d.getLineNumber(), d.getColumnNumber(), d.getSource(), d.getMessage(null));
+        }
+  
+        return b;
+      } finally {
+        fileManager.close();
       }
-      JavaCompiler.CompilationTask task = compiler.getTask(null, fileManager, diagnostics, optionsL, null, compilationUnits);
-      task.call();
     } finally {
-      fileManager.close();
+      Thread.currentThread().setContextClassLoader(cl0);
     }
   }
 }
